+++ SimulatorFileIO/Joints/RotationalJoint_Base.cs
@@ -11,10 +11,10 @@ public class RotationalJoint_Base : SkeletalJoint_Base
     public BXDVector3 childNormal;
     public BXDVector3 parentBase;
     public BXDVector3 childBase;
-    public double currentAngularPosition;
+    public float currentAngularPosition;
     public bool hasAngularLimit;
-    public double angularLimitLow;
-    public double angularLimitHigh;
+    public float angularLimitLow;
+    public float angularLimitHigh;
 
     public override SkeletalJointType getJointType()
     {
@@ -47,16 +47,16 @@ public class RotationalJoint_Base : SkeletalJoint_Base
 
     protected override void readJoint(System.IO.BinaryReader reader)
     {
-        parentBase = new BXDVector3(reader.ReadDouble(), reader.ReadDouble(), reader.ReadDouble());
-        parentNormal = new BXDVector3(reader.ReadDouble(), reader.ReadDouble(), reader.ReadDouble());
-        childBase = new BXDVector3(reader.ReadDouble(), reader.ReadDouble(), reader.ReadDouble());
-        childNormal = new BXDVector3(reader.ReadDouble(), reader.ReadDouble(), reader.ReadDouble());
+        parentBase = new BXDVector3(reader.ReadSingle(), reader.ReadSingle(), reader.ReadSingle());
+        parentNormal = new BXDVector3(reader.ReadSingle(), reader.ReadSingle(), reader.ReadSingle());
+        childBase = new BXDVector3(reader.ReadSingle(), reader.ReadSingle(), reader.ReadSingle());
+        childNormal = new BXDVector3(reader.ReadSingle(), reader.ReadSingle(), reader.ReadSingle());
 
         hasAngularLimit = (reader.ReadByte() & 1) == 1;
         if (hasAngularLimit)
         {
-            angularLimitLow = reader.ReadDouble();
-            angularLimitHigh = reader.ReadDouble();
+            angularLimitLow = reader.ReadSingle();
+            angularLimitHigh = reader.ReadSingle();
         }
     }
 }
